Extract munging and validation logic to DataTypeHandling module
The logic is extracted to lib/puppet/resource_api/data_type_handling.rb from
lib/puppet/resource_api.rb

This includes `mungify`, `mungify_core`, `try_mungify`, `validate`,
`try_validate`, `parse_puppet_type`, `validate_ensure`,
`ambiguous_error_msg`, `boolean_munge` methods.

unpack_strings parameter is added to mungify which allows migration
from legacy types.

The expose API is backwards compatibile.
